[WIP-ADD] Multiple objectives

+++ src/main/java/ar/edu/itba/montu/interfaces/KingdomObjective.java
@@ -6,8 +6,8 @@ import ar.edu.itba.montu.abstraction.LocatableAgent;
 import ar.edu.itba.montu.war.kingdom.Kingdom;
 
 public interface KingdomObjective extends Comparable<KingdomObjective> {
-	int priority();
-	void alterPriority(int priority);
+	double priority();
+	void alterPriority(double priority);
 	boolean involves(final LocatableAgent agent);
 	
 	List<Objective> translate();
